Updated shader code and texture shader to work with materials. This means that materials coded as specific colors and textures can even be overlapped during rendering

+++ Beon/headers/camera.hpp
@@ -101,7 +101,7 @@ public:
     }
 
     // Processes input received from any keyboard-like input system. Accepts input parameter in the form of camera defined ENUM (to abstract it from windowing systems)
-    void ProcessKeyboard(Camera_Movement direction, float deltaTime)
+    void ProcessKeyboard(Camera_Movement direction, double deltaTime)
     {
         float velocity = MovementSpeed * deltaTime;
         if (direction == FORWARD)
